feat: complete modules task
